pi-cloud-server/requirements.txt
# Pi Cloud Server - Python Dependencies
# =====================================

# Core Framework
Flask==3.0.0
Werkzeug==3.0.1

# Database
SQLAlchemy==2.0.23
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9

# Database Migrations
Flask-Migrate==4.0.5
alembic==1.13.0

# Environment Variables
python-dotenv==1.0.0

# Production Server
gunicorn==21.2.0

# Utilities
python-magic==0.4.27

# Development (optional)
# flask-debugtoolbar==0.14.1
